+++ dart/geodata/CHANGELOG.md
@@ -7,6 +7,9 @@
 🛠 Refactoring:
 * [Change some external dependency version requirements to any - at least for http package #247](https://github.com/navibyte/geospatial/issues/247)
 
+🛠 Maintenance:
+* [Update very_good_analysis version to 6.0.0 #249](https://github.com/navibyte/geospatial/issues/249)
+
 ## 1.2.1
 
 Small documentation link fixes.